Split out server logging
Split out server logging into logger.py since pbench/__init__.py
has gotten too big.
