+++ server/bin/pbench-backup-tarballs.py
@@ -7,8 +7,9 @@ import shutil
 import sys
 import tempfile
 
-from pbench import PbenchConfig, get_pbench_logger
+from pbench import PbenchConfig
 from pbench.exception import BadConfig
+from pbench.logger import get_pbench_logger
 from pbench.report import Report
 from pbench.utils import quarantine, rename_tb_link, md5sum
 from s3backup import NoSuchKey, S3Config, Status
